refactor: move colors.js to constants
