+++ src/components/Header/Button.js
@@ -28,6 +28,8 @@ const Button = ({ onClick }) => (
   </ButtonLayout>
 )
 
-Button.propTypes = {}
+Button.propTypes = {
+  onClick: PropTypes.func,
+}
 
 export default Button
