+++ src/components/Header/Logo.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import { WHITE } from '../../assets/colors'
+import { WHITE } from '../../constants/colors'
 
 const LogoLayout = styled.div`
   display: inline-block;
