+++ src/components/ResultPage/List/Category.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import { WHITE, MEDIUM_GREY } from '../../../assets/colors'
+import { WHITE, MEDIUM_GREY } from '../../../constants/colors'
 
 const CategoryLayout = styled.div`
   display: flex;
