+++ src/components/Sidebar/DatePicker.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import dateFns from 'date-fns'
